Move attr tests to their own file
